pyproject.toml
[tool.poetry]
name = "ds-auto-insights"
version = "0.1.0"
description = ""
authors = ["togunbiyi1 <olatunjiogunbiyi@googlemail.com>"]
readme = "README.md"
packages = [
  { include = "ds_auto_insights" }
]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.3.1"
fastapi = "^0.116.1"
openai = "^1.97.1"
python-multipart = "^0.0.20"
uvicorn = "^0.35.0"
python-dotenv = "^1.1.1"
streamlit = "^1.47.1"
tabulate = "^0.9.0"
langchain = "^0.3.27"
langchain-openai = "^0.3.32"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
